Polyfill MessageChannel via custom worker entrypoint
