+++ astro.config.mjs
@@ -18,6 +18,9 @@ export default defineConfig({
     platformProxy: {
       enabled: true,
     },
+    workerEntryPoint: {
+      path: "./src/worker.ts",
+    },
   }),
 
   vite: {
